renaming of the entry files
