+++ packages/edge-currency-bitcoin/package.json
@@ -26,7 +26,7 @@
   "main": "./index.js",
   "module": "./lib/index.js",
   "scripts": {
-    "build": "rimraf lib && sucrase ./src -q -d ./lib -t flow && webpack",
+    "build": "rimraf lib && sucrase ./src -q -d ./lib -t flow && webpack && mv ./lib/react-native.js ./lib/react-native-io.js",
     "flow": "flow",
     "format": "import-sort -l --write '*.js' 'src/**/*.js' 'test/**/*.js'; prettier-eslint --write '*.js' 'src/**/*.js' 'test/**/*.js'",
     "lint": "eslint '*.js' 'src/**/*.js' 'test/**/*.js'",
